Fix usage of n_chunks in Kraskov-OpenCL estimator
Remove the setting of nchunkspergpu to default 1 if this option is not
found in the opts structure. This is not needed because n_chunks is
passed as an argument to the function.

The bug resulted from merging to development branches, where one
branch fixed the setting of default values using the opts structure
and the other was for developing parallelization over chunks with the
GPU.