Fix bug in setting the default for n_permutations
The function did not handle the setting of n_permutations correctly. The
function should re-use the number of surrogates used for min_stats. If
this table does not exist, i.e., is None, trying to infer its first
dimension results in a TypeError because None can not be indexed. Thus
a TypeError needs to be caught if the table does not exist and we use
the default of 500 instead.