Add new class for single process analysis
Add a class Single_process_analysis on the same level as
Network_inference.  Make PID-estimation a child class of
Single_process_analysis instead of Network_inference. The
way the latter class is initialised doesn't suit the PID
analysis. In particular, the setting of a fixed min and max
lag does not work for PID analysis because we'll most likely
want individual lags for each process (this would require a
new class instance for each process, which isn't what we want).

Rename Single_process_storage to Active_information_storage.
This is more consistent with class naming conventions in the
rest of the toolbox.

Change class names in unit tests.