Bugfix: test for variable size using a static variable name
Existing code used a static name ('var1') to access a variable
in the data dictionary for serial estimation. If an estimator
doesn't use var1 as argument this breaks (e.g. for the PID
estimators). Fix this by using the variable name provided in
the slice_var list, which holds variables that need to be cut
up for serial estimation.