Refactor: change name of surrogate method in stats
Use _get_surrogates to make it clearer that surrogates are not
created in the stats module but by the data class. The method
just serves to choose the correct type of surrogate creation
depending on the number of replications and the user settings.
This is also made explicit in the documentation of the method.