Fix testing for int by comparing with np.integer
Test if arrays contain ints by testing wether they are subclasses of
np.integer instead of testing whether they are subclasses of np.int64.
The latter limits the program to 64-bit architectures and ints of type
int64. The class np.integer is a superclass of int32 and int64 such
that it can be used to test if an int is of any of the two types.

Note that int32 is not a subclass of int64 and Python uses the default
int type, which depends on the architecture (32 or 64 bit). Thus, we
can not compare against np.int64 as this will fail on 32-bit machines.
Also, we can not compare against np.int_ because this just represents
the default type on the system. We may get problems if the user feeds
in data that is not the default int but nevertheless an int (e.g., if
s.o. uses int64 on a 32-bit architecture).