Unify user input checks when creating an instance
Unify the way user input is checked for correctness when creating
an instance of Active_information_storage and Multivariate_te.
Check input when creating the object not when running the analysis.
Input should be checked immediately when creating the object, not
when running the first analysis.

Write unit tests. Unift tests for both classes.