Fix errors in class structure and JIDT properties
Remove estimate method from JidtEstimator, the class should not be
instantiated. If the abstract method estimate is not implemented an
error is raised upon instantiation.

Make JidtGaussian inherit from JidtEstimator. Gaussian estimators don't
use the same properties as the Kraskov estimators. Add documentation for
this.

Fix checking for empty conditionals (None) in CMI estimators.

Remove comparison of tau and history, tau can be larger than the history
(i.e., the embedding dimension). This was an error. Fix respective unit
test.

Let JIDT add the noise, it is creating a new array anyway so we don't
have to copy data twice.