Fix bug, add assertions
Fix bug in getting the size of the minimum stats surrogate table. If
table is None, Python will throw an AttributeError when
max_stats_sequential tries to access the size of the table. The try/
except block should thus catch an AttributeError instead of a TypeError.

Asset that the set of selected sources is not empty for tests that make
use of this set.