Fix passing of options dictionary to stats functions
Do not pass options for statistical testing as a separate argument
because the opts ditctionary is passed as an attribute of the analysis
setup anyways. Add defaults to the opts dict if a parameter was not set
so the analysis performed is documented in the analysis object.

Add documentation and fix function calls such that they do not pass the
options dictionary as a separate argument. Fix unit tests.

Make FDR-correction optional and handle parameters for FDR correction
like parameters for all other tests.