Bugfix in comparison of values against surrogate distribution
Fix comparison of statistic against surrogate distribution. Statistic
needs to be truly bigger/smaller than surrogates to be considered
significant. Before, equal surrogates did not enter the numerator when
calculating the p-value, resulting in smaller p-values. In the extreme
case of all surrogates being equal to the test statistic, the statistic
would be considered significant even though it never exceeded the surrogates.