Pass all analysis parameters in options dictionary
Pass all analysis parameters (lags, taus, stats options, forced
conditional variables, etc.) in a single dictionary. The dictionary
is passed to the class methods performing the analysis. In earlier
versions, some parameters were passed to the classes upon
instantiation.

This moves all parameters into one place for more consistency and
usability. The analysis_options dict is added to the algorithm class
as an attribute, and is returned to the user after the calculation
has finished. All parameters used in the estimation are added to the
options dictionary, including default parameters set by the algorithm
to document all parameters used for analysis.