Remove defaults from statistical tests
Remove default values for statistical tests such that these defaults
must be specified in the calling functions. The calling functions
specify defaults and add them to the options dictionary. The options
dict is an attribute of the algorithm class and is thus constantly
updated throughout the execution of the algorithm. This way, all
analysis options get documented by adding and returned to the user.

Defaults concern the statistical tests themselves as well as the
generation of surrogates in the Data() class.