Add sanity check to FDR correction
Add sanity check to FDR correction that makes sure the FDR threshold
can be reached given the smallest possible p-value in the data. The
smallest possible p-value is 1 / no. permutations used for sequential
max stats (the p-values calculated in this test are added to the
results dict). Issue a warning and abort if this smallest theoretically
possible p-value is already larger than the FDR threshold.