Add checks for no. data points for Kraskov estimators
Add checks for no. data points for Kraskov estimators. If the number
of points - 1  is equal to the number of nearest neighbours
('kraskov_k'), the estimated value will be a constant. If the number
is smaller, no estimation is possible. Add checks similar to those
performed internally by JIDT. Take the Theiler correction and source-
target-delays into account.