Compute CMI only once for between tests
Compute CMI only once for between tests. The CMI in the union network
needs to be computed only once for the between test. To generate
surrogate data, the CMI is shuffled between groups before the diff
of the mean gets recomputed. This doesn't require the re-computation
of the CMI for each surrogate run. This results in a significant
speedup. Remove console output for surrogate creation, we don't need
this anymore, since the surrogate creation is done after a few seconds.

Additionally return a comparison between mean TE values in the two
groups of the between test. The sign of the difference of the raw values
may be misleading if one or both values are negative.