Return absolute difference in TE values
Return absolute difference in TE values and the direction of the
effect (A>B vs. A<=B). Evaluating the sign of the raw difference
may be misleading for negative TE values which may happen due to
estimator bias. Absolute difference and direction of effect are
reported as two separate entries in the results dict.

Add the number of subjects to the settings dict for between tests.