Fix bug in enforcing conditionals
Fix a bug in the method that enforces conditionals in the conditioning
set for AIS estimation and network inference using bivariate or multivariate
TE. The method assumed conditioning variables were provided using absolute
indices in time, but they are actually provided using lags. Fix this in
the method and add unit tests for this functionality.