Fix bug in discrete estimator
Fix bug in discrete estimator. The discretisation method checks if data
are discrete if no discretisation is requested when using discrete
estimators. One of the checks the method performs is a test if the
symbols in the provided time series match the provided alphabet sizes.
The method uses min and max together with assertions to realise this.
However, min and max when applied to multidimentional arrays return 1D
arrays of minima and maxima along the first axis of the array. If these
arrays are tested for some condition within an assert statement, the
assert statement receives something other than the expected boolean
value as first argument. The fix uses numpy min and max methods, which
always take the minimum and maximum over the whole array, not only a
single axis.