Allow group stats for results from MI network inference
Allow group stats for  results from MI network inference. Handle empty
conditioning sets when estimating the MI/TE for single links. This happens
if there are no other sources to condition on and no variables in the
target's past (which is always the case when inferring links unsing MI).

Add None as return value for _get_realisations in the Data class.
This is consistent with the estimators which can handle None as an input
for a variable but not empty arrays.

Minor corrections in class docstring.