Allow 'one' as test tail
Allow 'one' as test tail (in addition to 'one_bigger'). This is more in
line with conventions. In most cases we want 'one_bigger', i.e., we
want to test whether the test statistic is bigger than the surrogates
(e.g., when testing TE estimates against surrogates). 'one' as a
keyword is more what one would expect from a test that mainly does an
estimate-against-surrogates-comparison.