Fix bug in OpenCL code execution on AMD cards
Fix bug in OpenCL code execution on AMD cards. The size of the point set
has to be a multiple of 256. This means we have to have a multiple of 64
number of points since one point has 4 Byte (float32). Pad data such that
we get the desired number of points. When padding data note that all
chunks have to have the same number of points, meaning we can't just add
the number of missing points at the end of the whole point set. Instead,
try to pad data by adding/taking away up to 10 points. If this is not
sufficient start adding whole chunks with garbage data and repeat the
adding/taking away of up to 10 points. This is repeated until a sufficient
number of points is reached. Perform padding for AMD devices only.

Change unit tests to reflect changes. In particular: change neighbour search
unit tests such that point sets have a size of 64 or some multiple of 64.
Add unit tests for padding function.