Fix bug in bivariate MI/TE estimation and min stats
Fix bug in bivariate MI/TE estimation. Conditioning was not performed
correctly. Find non-uniform embedding separately for each link. This
requires conditioning on the target's past for TE and conditioning on
all variables already selected for this link.

Fix bug in minimum statistics: conditional for surrogate creation should
not contain the minimum candidate. Otherwise the minimum candidate's CMI
is calculated with a conditional that is smaller by one dimension
compared to the conditional used for surrogate creation.

Add unit tests and change documentation.