Fix copying of settings dict for estimators and PID
Fix copying of settings dict for estimators and PID. Settings dict must
be copied, otherwise external changes in the dictionary result in
changed behavior of the estimators in the next call to estimate(). Make
sure that setting default values happens on the settings dict passed as
an input such that the settings in the calling function get updated.
This serves as a log for algo classes.

Add handling of 0 as noise level for OpenCL estimator. Code crashed if
noise was set to 0.