Fix bug in local values calculation
Fix bug in local values calculation. Bug was introduced by fixing a bug
in the setting of estimator settings. The estimator settings dictionary
is now copied when it is added to an estimator instance. Before it
wasn't copied which led to unindented behavior, i.e., changes in the
settings dictionary had an effect on existing estimator instances. Now
this sideeffect is fixed, however, calculation of local values returned
by the algos exploited this bug. This commit fixes the broken
calculation of local values by initialising two estimators (for local and
average values) if the user requested the calculation of local values.
Move code for estimator initialisation into helper method shared by all
agorithm classes.