Set up Sphinx for publication on gh-pages
Set up Sphinx for publication on gh-pages. Keep documentation in the
docs folder. For html documentation to be published on gh-pages, add
a .nojekyll file and an index.html that points to the html build dir.
Documentation can no be updated by running 'make html' in the docs
folder, committing, and pushing to GitHub.

Fix typos.