Fix bug in max stats sequential
Fix bug in max stats sequential. Set default to n_perm_min_stat
setting instead of size of surrogate table. If for the previous target,
only one candidate was included, no surrogate table is created, and the
n_perm for max stats sequential is set to the default of 500. If for the
next target more candidates are included and a proper pruning step is
performed, a surrogate tabel with a differnet n_perm may be created. This
results in two targets with unequal analysis settings, which causes an
error when combining results in the network inference algorithm. This is