Add PEP8 and documentation for Goettingen PID
Add PEP8 and documentation for Goettingen PID. Make PrettyTable import
optional and issue a warning similar to the ECOS estimator for the Tartu
PID. Otherwise, any analysis crashes because it tries to import somthing
that is optional and thus probably not installed. Rename current PID
estimators to BivariatePID and new estimator to MultivariatePID to be
consistent with MI and TE class names.

Fix import error in multivariate PID estimator.