Fix calculation of single-link estimates for multivariate algorithms
Fix calculation of single-link estimates for multivariate algorithms.
Conditioning in calculating output values used only target past states
in multivariate algorithms, instead of the past of all other selected
sources as well. Note that this values is not used by the algorithm, but
is provided in the output for completeness only.