Cast return values of JIDT estimators to float
Cast return values of JIDT estimators to float. Without this, return
values may be JPype types. This is inconvenient, for example, when
saving IDTxl results to disk and loading them in a new session, JPype
types require that a JVM is running to load the results. If the JVM
does not run, pickle fails to load the results with a hard to
understand error message.