Ensure network_analysis handles partially empty networks
If single links have no significant inputs, i.e., no source variables
are selected, no Omnibus test is performed. If the user did not specify
the no. permutations for the Omnibus test, this parameter is not set.
Ensure that this does not lead to problems when combining results across
targets and one target has no Ombinus test settings, while the other one
does.